move relevant functions to gadeable.cpp and extract string_utility functions
